+++ src/app/layout.server.tsx
@@ -1,0 +1,5 @@
+import { metadata } from "./metadata";
+import RootLayout from "./layoutClient";
+
+export { metadata };
+export default RootLayout;
