+++ src/app/metadata.ts
@@ -1,5 +1,5 @@
 import type { Metadata } from "next";
 
 export const metadata: Metadata = {
-  title: "KiiChain",
+  title: "KiiChain Explorer",
 };
